[#33] Move split tests accordingly
