+++ src/test/java/de/miesner/claus/bingo/table/RowTest.java
@@ -4,7 +4,6 @@ import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
-import static de.miesner.claus.bingo.util.latex.Expression.MAX_CHARS_PER_ROW;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class RowTest {
@@ -62,73 +61,5 @@ class RowTest {
     this.row = new Row(List.of("one"));
     var equalRow = new Row(List.of("one"));
     assertThat(row.equals(equalRow)).isTrue();
-  }
-
-  @Test
-  void testSplitNoSplitNeeded() {
-    String word = "one";
-    String expected = "one";
-    assertThat(word.length()).as("The test is valid.").isLessThanOrEqualTo(MAX_CHARS_PER_ROW);
-
-    assertThat(row.split(word, MAX_CHARS_PER_ROW)).isEqualTo(expected);
-  }
-
-  @Test
-  void testSplitWordHasMaxLength() {
-    String word = "abbreviation";
-    String expected = "abbreviation";
-    assertThat(word.length()).as("The test is valid.").isEqualTo(MAX_CHARS_PER_ROW);
-
-    assertThat(row.split(word, MAX_CHARS_PER_ROW)).isEqualTo(expected);
-  }
-
-  @Test
-  void testSplitWordTooLong() {
-    String word = "jackhammering";
-    String expected = "jackhammer- ing";
-    assertThat(12).as("The test is valid.").isEqualTo(MAX_CHARS_PER_ROW);
-
-    assertThat(row.split(word, MAX_CHARS_PER_ROW)).isEqualTo(expected);
-  }
-
-  @Test
-  void testSplitTwoWordsFit() {
-    String word = "jacks hammer";
-    String expected = "jacks hammer";
-    assertThat(word.length()).as("The test is valid.").isLessThanOrEqualTo(MAX_CHARS_PER_ROW);
-
-    assertThat(row.split(word, MAX_CHARS_PER_ROW)).isEqualTo(expected);
-  }
-
-  @Test
-  void testSplitTwoWordsDontFitButSpaceIsAlreadyThere() {
-    String word = "jack's hammering";
-    String expected = "jack's hammering";
-    assertThat(word.length()).as("The test is valid.").isGreaterThan(MAX_CHARS_PER_ROW);
-    assertThat(word.indexOf(" ")).as("The test is valid.").isLessThanOrEqualTo(MAX_CHARS_PER_ROW - 1);
-
-    assertThat(row.split(word, MAX_CHARS_PER_ROW)).isEqualTo(expected);
-  }
-
-  @Test
-  void testSplitMultipleWordsSpacesPresent() {
-    String word = "jacks hammer hammers";
-    String expected = "jacks hammer hammers";
-    assertThat(word.length()).as("The test is valid.").isGreaterThan(MAX_CHARS_PER_ROW);
-    int indexFirstSpace = word.indexOf(" ");
-    int indexSecondSpace = word.indexOf(" ", indexFirstSpace + 1);
-    assertThat(indexFirstSpace).as("The test is valid.").isLessThanOrEqualTo(MAX_CHARS_PER_ROW);
-    assertThat(indexSecondSpace - indexFirstSpace).as("The test is valid.").isLessThanOrEqualTo(MAX_CHARS_PER_ROW);
-
-    assertThat(row.split(word, MAX_CHARS_PER_ROW)).isEqualTo(expected);
-  }
-
-  @Test
-  void testSplitMultipleWordsNeedToBeSplit() {
-    String word = "jack's jackhammering massively";
-    String expected = "jacks's jackhammer- ing massively";
-    assertThat(12).as("The test is valid.").isEqualTo(MAX_CHARS_PER_ROW);
-
-    assertThat(row.split(word, MAX_CHARS_PER_ROW)).isEqualTo(expected);
   }
 }
